KTOR-3429 add DAOFacade interface
